Refactor to allow for new features
