+++ data/sample-repos.json
@@ -1,13 +1,13 @@
-[
+{
+  "repos": [
     {
-        "name": "Sample Project Name",
-        "ssh": "git@github.com:Sample/sample-project-name.git",
-        "location": "/ptah/on/local/disc/to/sample-project-name",
-        "npm": true, // optional flag
-        "in_go_path": true, // optional flag, information only
-        "projects": [
-            ".",
-            "subpackage/path"
-        ]
+      "name": "Tutorial Web App Walkthroughs",
+      "ssh": "git@github.com:integr8ly/tutorial-web-app-walkthroughs.git",
+      "npm": true,
+      "location": "/home/jimfitz/SourceClear/repos/github.com/integr8ly/tutorial-web-app-walkthroughs",
+      "projects": [
+        "."
+      ]
     }
-]
+  ]
+}
